Add specific types for sleeve tasks
* Move LocationName enum to overall enums file
* Add LocationName enum to those provided to player
* Add specific SleeveTask types for various types of sleeve work
* remove trailing s on CompanyPosNames enum
