+++ src/utils/enums.ts
@@ -1,4 +1,4 @@
-export enum CompanyPosNames {
+export enum CompanyPosName {
   sw0 = "Software Engineering Intern",
   sw1 = "Junior Software Engineer",
   sw2 = "Senior Software Engineer",
@@ -75,4 +75,71 @@ export enum GymType {
   defense = "def",
   dexterity = "dex",
   agility = "agi",
+}
+
+/** Names of all locations */
+export enum LocationName {
+  AevumAeroCorp = "AeroCorp",
+  AevumBachmanAndAssociates = "Bachman & Associates",
+  AevumClarkeIncorporated = "Clarke Incorporated",
+  AevumCrushFitnessGym = "Crush Fitness Gym",
+  AevumECorp = "ECorp",
+  AevumFulcrumTechnologies = "Fulcrum Technologies",
+  AevumGalacticCybersystems = "Galactic Cybersystems",
+  AevumNetLinkTechnologies = "NetLink Technologies",
+  AevumPolice = "Aevum Police Headquarters",
+  AevumRhoConstruction = "Rho Construction",
+  AevumSnapFitnessGym = "Snap Fitness Gym",
+  AevumSummitUniversity = "Summit University",
+  AevumWatchdogSecurity = "Watchdog Security",
+  AevumCasino = "Iker Molina Casino",
+
+  ChongqingKuaiGongInternational = "KuaiGong International",
+  ChongqingSolarisSpaceSystems = "Solaris Space Systems",
+  ChongqingChurchOfTheMachineGod = "Church of the Machine God",
+
+  Sector12AlphaEnterprises = "Alpha Enterprises",
+  Sector12BladeIndustries = "Blade Industries",
+  Sector12CIA = "Central Intelligence Agency",
+  Sector12CarmichaelSecurity = "Carmichael Security",
+  Sector12CityHall = "Sector-12 City Hall",
+  Sector12DeltaOne = "DeltaOne",
+  Sector12FoodNStuff = "FoodNStuff",
+  Sector12FourSigma = "Four Sigma",
+  Sector12IcarusMicrosystems = "Icarus Microsystems",
+  Sector12IronGym = "Iron Gym",
+  Sector12JoesGuns = "Joe's Guns",
+  Sector12MegaCorp = "MegaCorp",
+  Sector12NSA = "National Security Agency",
+  Sector12PowerhouseGym = "Powerhouse Gym",
+  Sector12RothmanUniversity = "Rothman University",
+  Sector12UniversalEnergy = "Universal Energy",
+
+  NewTokyoDefComm = "DefComm",
+  NewTokyoGlobalPharmaceuticals = "Global Pharmaceuticals",
+  NewTokyoNoodleBar = "Noodle Bar",
+  NewTokyoVitaLife = "VitaLife",
+  NewTokyoArcade = "Arcade",
+
+  IshimaNovaMedical = "Nova Medical",
+  IshimaOmegaSoftware = "Omega Software",
+  IshimaStormTechnologies = "Storm Technologies",
+  IshimaGlitch = "0x6C1",
+
+  VolhavenCompuTek = "CompuTek",
+  VolhavenHeliosLabs = "Helios Labs",
+  VolhavenLexoCorp = "LexoCorp",
+  VolhavenMilleniumFitnessGym = "Millenium Fitness Gym",
+  VolhavenNWO = "NWO",
+  VolhavenOmniTekIncorporated = "OmniTek Incorporated",
+  VolhavenOmniaCybersystems = "Omnia Cybersystems",
+  VolhavenSysCoreSecurities = "SysCore Securities",
+  VolhavenZBInstituteOfTechnology = "ZB Institute of Technology",
+
+  Hospital = "Hospital",
+  Slums = "The Slums",
+  TravelAgency = "Travel Agency",
+  WorldStockExchange = "World Stock Exchange",
+
+  Void = "The Void",
 }
